fixed a bunch of non critical warnings, and commented some unfinished code

+++ src/path.cpp
@@ -49,7 +49,7 @@ void Path::updatePath(std::vector<Node3D> nodePath) {
   path.header.stamp = ros::Time::now();
   int k = 0;
 
-  for (int i = 0; i < nodePath.size(); ++i) {
+  for (size_t i = 0; i < nodePath.size(); ++i) {
     addSegment(nodePath[i]);
     addNode(nodePath[i], k);
     k++;
